Api Branch version 02

+++ src/main/java/com/iset/bp/web/ContactController.java
@@ -5,12 +5,10 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -43,7 +41,10 @@ public class ContactController {
 	
 	@DeleteMapping("/contacts/{id}")
 	public void SupprimerContact(@PathVariable int id) {
-		 contactRep.deleteById(id);
+		Optional<Contact> contact = contactRep.findById(id);
+		if (contact.isPresent()) { 
+			contactRep.deleteById(id);
+	    }else throw new RuntimeException("Contact introuvable on ne peut pas le supprimer !");
 	}
 	
 	@RequestMapping(value="/contacts",method = RequestMethod.POST)
